try not to assume size_t == unsigned long

+++ src/scopmath/sparse_thread.c
@@ -127,7 +127,7 @@ static void prmat(SparseObj* so);
 static void initeqn(SparseObj* so, unsigned maxeqn);
 static void free_elm(SparseObj* so);
 static Elm* getelm(SparseObj* so,unsigned row, unsigned col, Elm* new);
-static void create_coef_list(SparseObj* so, int n, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, unsigned long iml);
+static void create_coef_list(SparseObj* so, int n, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, size_t iml);
 static void init_coef_list(SparseObj* so);
 static void init_minorder(SparseObj* so);
 static void increase_order(SparseObj* so, unsigned row);
@@ -153,7 +153,7 @@ ensures all elements needed are present */
 
 #define s_(arg) *p[s[arg]]
 #define d_(arg) *p[d[arg]]
-int sparse_thread(void** v, int n, int* s, int* d, double** p, double* t, double dt, sparse_fptr fun, int linflag, Datum* ppvar, Datum* thread, NrnThread *nt, Memb_list* ml, unsigned long iml) {
+int sparse_thread(void** v, int n, int* s, int* d, double** p, double* t, double dt, sparse_fptr fun, int linflag, Datum* ppvar, Datum* thread, NrnThread *nt, Memb_list* ml, size_t iml) {
 	int i, j, ierr;
 	double err;
 	SparseObj* so;
@@ -197,7 +197,7 @@ if (!linflag && s_(i-1) < 0.) { s_(i-1) = 0.; }
 }
 
 /* for solving ax=b */
-int _cvode_sparse_thread(void** v, int n, int* x, double** p, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, unsigned long iml) {
+int _cvode_sparse_thread(void** v, int n, int* x, double** p, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, size_t iml) {
 #define x_(arg) *p[x[arg]]
 	int i, j, ierr;
 	SparseObj* so;
@@ -475,7 +475,7 @@ double* _nrn_thread_getelm(SparseObj* so, int row, int col) {
 	return &el->value;
 }
 
-static void create_coef_list(SparseObj* so, int n, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, unsigned long iml) {
+static void create_coef_list(SparseObj* so, int n, sparse_fptr fun, Datum* ppvar, Datum* thread, NrnThread* nt, Memb_list* ml, size_t iml) {
 	initeqn(so, (unsigned)n);
 	so->phase = 1;
 	so->ngetcall = 0;
